Extracted the string manipulation functions to useful_functions.py Working on getting a base computation class for sGTM to allow for label mappings to work
